Automate doc, tests, lint, packaging, releasing and publishing
Ignore IDEA dist folder

Wip

v1.0.0

release v1.0.0-alpha.3

release v1.0.0

Automate doc generation (TOC + API) and releasing/publishing steps

release v1.0.1-beta.0

release v1.0.1-beta.1

release v1.0.1-beta.2

Fix couple issues

release v1.0.1

Publish API file on NPM (doc) as well
