+++ .gitignore
@@ -120,9 +120,8 @@ typings/
 
 # Yarn Integrity file
 .yarn-integrity
+yarn-error.log
 
 # dotenv environment variables file
 .env
 
-# next.js build output
-.next
