+++ .idea/utils.iml
@@ -1,12 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <module type="WEB_MODULE" version="4">
-    <component name="NewModuleRootManager">
-        <content url="file://$MODULE_DIR$">
-            <excludeFolder url="file://$MODULE_DIR$/.tmp" />
-            <excludeFolder url="file://$MODULE_DIR$/temp" />
-            <excludeFolder url="file://$MODULE_DIR$/tmp" />
-        </content>
-        <orderEntry type="inheritedJdk" />
-        <orderEntry type="sourceFolder" forTests="false" />
-    </component>
+  <component name="NewModuleRootManager">
+    <content url="file://$MODULE_DIR$">
+      <excludeFolder url="file://$MODULE_DIR$/.tmp" />
+      <excludeFolder url="file://$MODULE_DIR$/dist" />
+      <excludeFolder url="file://$MODULE_DIR$/temp" />
+      <excludeFolder url="file://$MODULE_DIR$/tmp" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+  </component>
 </module>
