--- index.js
@@ -1,9 +1,0 @@
-const convertLineBreaks = require('./lib/convertLineBreaks');
-const isBrowser = require('./lib/isBrowser');
-const browserRedirect = require('./lib/browserRedirect');
-
-module.exports = {
-  convertLineBreaks,
-  isBrowser,
-  browserRedirect,
-};
