+++ lib/browserRedirect.js
@@ -1,12 +1,16 @@
+"use strict";
+
 /**
  * Redirect that handles the HTML <a> target
  * to redirect the current frame or open a new tab.
  *
- * @param redirect.url
- * @param redirect.target
+ * @param url
+ * @param target
  */
-const browserRedirect = (redirect) => {
-  const { url, target } = redirect;
+var browserRedirect = function browserRedirect(_ref) {
+  var url = _ref.url,
+      target = _ref.target;
+
   if (target === '_blank') {
     window.open(url, target);
   } else {
