+++ lib/convertLineBreaks.js
@@ -1,3 +1,7 @@
+"use strict";
+
+function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterator === "symbol") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }; } return _typeof(obj); }
+
 /**
  * Replace line breaks in a string by a HTML line break (<br>)
  *
@@ -5,9 +9,11 @@
  * @param replacer
  * @returns {string}
  */
-const convertLineBreaks = (str, replacer = '<br>') => {
+var convertLineBreaks = function convertLineBreaks(str) {
+  var replacer = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : '<br>';
+
   if (typeof str !== 'string') {
-    throw Error(`convertLineBreaks waiting for string but receive a ${typeof str}`);
+    throw Error("convertLineBreaks waiting for string but receive a ".concat(_typeof(str)));
   }
 
   return str.replace(/\n/g, replacer);
