+++ lib/isBrowser.js
@@ -1,8 +1,12 @@
+"use strict";
+
 /**
  * Checks whether the current runtime is a browser
  *
  * @returns {boolean}
  */
-const isBrowser = () => typeof window !== 'undefined';
+var isBrowser = function isBrowser() {
+  return typeof window !== 'undefined';
+};
 
 module.exports = isBrowser;
