+++ package.json
@@ -1,30 +1,45 @@
 {
   "name": "@unly/utils",
-  "version": "1.0.0-alpha.1",
+  "version": "1.0.1",
   "description": "Common utils",
-  "main": "index.js",
+  "repository": "https://github.com/UnlyEd/utils.git",
+  "homepage": "https://github.com/UnlyEd/utils",
   "author": "unlyEd",
   "license": "MIT",
-  "repository": "https://github.com/UnlyEd/utils.git",
   "bugs": {
     "url": "https://github.com/UnlyEd/utils.git/issues"
   },
-  "homepage": "https://github.com/UnlyEd/utils",
+  "main": "lib/index.js",
+  "files": [
+    "/lib",
+    "/API.md"
+  ],
+  "publishConfig": {
+    "registry": "https://registry.npmjs.org/",
+    "access": "public"
+  },
   "jest": {
     "verbose": true,
     "testEnvironment": "node"
   },
   "scripts": {
-    "lint": "eslint . --cache --fix",
-    "test": "NODE_ENV=test jest",
-    "test:watch": "NODE_ENV=test jest --watch",
-    "test:coverage": "NODE_ENV=test jest --coverage",
-    "build": "babel lib --out-dir dist",
-    "preversion": "yarn run lint && yarn run test",
+    "start": "cross-env-shell 'concurrently -p '{name}' -n 'lint,build,test' -c 'gray.bgWhite,yellow.bgBlue,green.bgWhite' \"yarn lint\" \"yarn build\" \"yarn test\"'",
+    "build": "babel src/ --out-dir lib/ --watch",
+    "build:once": "babel src/ --out-dir lib/",
+    "clean": "rm -rf lib/",
+    "lint": "esw src/ --fix --watch",
+    "lint:once": "eslint src/ --fix",
+    "preversion": "yarn lint:once && yarn test:once && yarn doc:toc && yarn run doc:api",
     "postversion": "git push && git push --tags",
-    "release:beta": "npm publish --tag beta --access=public",
-    "release:alpha": "npm publish --tag alpha --access=public",
-    "release:latest": "npm publish --tag latest --access=public",
+    "prepublishOnly": "yarn clean && yarn build:once && yarn publish:preview && cli-confirm \"Do you really want to release a new version? Please check the files that will be publicly released first.\"",
+    "publish:preview": "npm pack && tar -xvzf *.tgz && rm -rf package *.tgz",
+    "release": "yarn bump --prompt --commit --tag --push",
+    "releaseAndPublish": "yarn release && npm publish",
+    "doc:api": "jsdoc2md src/*.js > API.md",
+    "doc:toc": "yarn markdown-toc --maxdepth 4 -i README.md",
+    "test": "NODE_ENV=test jest --watch",
+    "test:once": "NODE_ENV=test jest",
+    "test:coverage": "NODE_ENV=test jest --coverage"
   },
   "devDependencies": {
     "@babel/cli": "7.2.3",
@@ -33,18 +48,16 @@
     "@babel/preset-react": "7.0.0",
     "babel-core": "7.0.0-bridge.0",
     "babel-jest": "23.6.0",
+    "cli-confirm": "1.0.1",
+    "concurrently": "4.1.0",
+    "cross-env": "5.2.0",
     "eslint": "4.19.1",
     "eslint-config-airbnb-base": "13.1.0",
     "eslint-plugin-import": "2.14",
-    "jest": "23.6.0"
-  },
-  "files": [
-    "dist",
-    "lib",
-    "index.js",
-    "README.md"
-  ],
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/"
+    "eslint-watch": "4.0.2",
+    "jest": "23.6.0",
+    "jsdoc-to-markdown": "4.0.1",
+    "markdown-toc": "1.2.0",
+    "version-bump-prompt": "4.2.2"
   }
 }
