+++ src/index.js
@@ -1,0 +1,9 @@
+const convertLineBreaks = require('./convertLineBreaks');
+const isBrowser = require('./isBrowser');
+const browserRedirect = require('./browserRedirect');
+
+module.exports = {
+  convertLineBreaks,
+  isBrowser,
+  browserRedirect,
+};
