+++ src/isBrowser.js
@@ -1,0 +1,8 @@
+/**
+ * Checks whether the current runtime is a browser
+ *
+ * @returns {boolean}
+ */
+const isBrowser = () => typeof window !== 'undefined';
+
+module.exports = isBrowser;
